Okay
Lourds

+++ CLUEDO/Assets/Cluedo/Scripts/Save/CE_DataPath.cs
@@ -5,25 +5,34 @@ using System.IO;
 public class CE_DataPath
 {
 
-    public static string DataPath => Path.Combine(Application.persistentDataPath, DataPathExtension);
-    static string DataPathExtension => "SaveCluedo.fdp";
-
-    public static string DataPathSecure => Path.Combine(Application.persistentDataPath,  DataPathSecureExtension);
-    static string DataPathSecureExtension => "Porn/ghnhuz.TOTORO";
+    
+    public static string DataPath => Path.Combine(FolderDataPath, DataPathExtension);
+    public static string FolderDataPath => Path.Combine(Application.persistentDataPath, currentUser);
+    static string DataPathExtension => "/SaveCluedo.fdp";
+    static string currentUser = "";
+    public static string DataPathSecure => Path.Combine(FolderDataPathSecure,  DataPathSecureExtension);
+    public static string FolderDataPathSecure => Path.Combine(Application.persistentDataPath,  currentUser);
+    static string DataPathSecureExtension => "/Porn/ghnhuz.TOTORO";
     
     static bool InitFolder()
     {
-        if (!Directory.Exists(DataPathSecure.Replace("/ghnhuz.TOTORO", "")))
-            Directory.CreateDirectory(DataPathSecure.Replace("/ghnhuz.TOTORO", ""));
+        Debug.Log(currentUser);
+        if (!Directory.Exists(FolderDataPath))
+            Directory.CreateDirectory(FolderDataPath);
 
-        return Directory.Exists(DataPathSecure.Replace("/ghnhuz.TOTORO", ""));
+        if (!Directory.Exists(FolderDataPathSecure))
+            Directory.CreateDirectory(FolderDataPathSecure);
+
+        return Directory.Exists(FolderDataPathSecure);
     }
     public static bool FileExist()
     {
+        if (InitFolder() && !File.Exists(DataPathSecure))
+        {
+            File.WriteAllText(DataPathSecure, "");
+        }
         if (!File.Exists(DataPath))
             File.WriteAllText(DataPath,"");
-        if (!InitFolder())
-            File.WriteAllText(DataPathSecure,"");
 
         return File.Exists(DataPath) && File.Exists(DataPathSecure);
     }
@@ -42,8 +51,11 @@ public class CE_DataPath
             File.WriteAllText(DataPathSecure, "");
         }
     }
-    public static bool IsSave()
+    public static bool IsSave(string _user)
     {
+        Debug.Log(_user);
+        currentUser = _user;
+        Debug.Log(currentUser);
         VerifSave();
         return File.ReadAllText(DataPath) != string.Empty;
     }
